--- diagnostics/finite_amplitude_wave_diag/coordinate_utils.py
@@ -1,26 +1,0 @@
-import numpy as np
-from hn2016_falwa.constant import P_GROUND, SCALE_HEIGHT
-
-
-def convert_pseudoheight_to_hPa(height_array):
-    """
-    Args:
-        height_array(np.array): pseudoheight in [m]
-
-    Returns:
-        np.array which contains pressure levels in [hPa]
-    """
-    p_array = P_GROUND * np.exp(- height_array / SCALE_HEIGHT)
-    return p_array
-
-
-def convert_hPa_to_pseudoheight(p_array):
-    """
-    Args:
-        height_array(np.array): pseudoheight in [m]
-
-    Returns:
-        np.array which contains pressure levels in [hPa]
-    """
-    height_array = - SCALE_HEIGHT * np.log(p_array / P_GROUND)
-    return height_array
