--- diagnostics/finite_amplitude_wave_diag/data_preprocessor.py
@@ -1,3 +1,0 @@
-
-
-
